bson: Add bson_validate_key().
Implement a bson_validate_key() function, that can be used by the
application developer to verify whether a BSON key is valid.

Depending on the context, this might mean disallowing dots, or
forbidding dollar signs at the beginning of the string.

Since this validation is context sensitive, it is left up to the
application developer to call this function at the appropriate places.