debian: Use dh_makeshlibs -V.
Without -V, the shlibs file will contain an unversioned dependency on
libmongo-client0 (or whatever the major is), which we do not
want. While we can guarantee that we remain backwards compatible as
long as the SONAME remains the same, there will be additions to the
library nevertheless, which do not exist in older versions.

Thus, the safest route is to depend on at least the version that's
currently present.