sync-cursor: Start at offset -1, instead of 0.
Instead of starting at offset 0, at the first element of the result
set, start at -1, so we'll need a mongo_sync_cursor_next() before we
can get the data.

This is so that the interface will be similar to bson_cursor_new(),
which already behaves this way, and because this way, we can safely
write code like the following:

  while (mongo_sync_cursor_next (cursor))
    {
      /* Do something interesting */
    }

Which is much clearer than having to put the while at the end.

The test cases were updated for the new behaviour, and new tests were
added to spot the differences between the old and new behaviours.