mongo-sync: Support passive secondaries.
Secondaries with priority 0 do not appear in ismaster's 'host' array,
but they're present in 'passives'. Since for reading, these passives
are perfectly fine, append them to our host set.

This means that passives are automatically discovered aswell, and the
library treats them as any other server. Which in turn makes the
connection pool work even when there's a master and passives.

No test cases were written, as testing this requires external setup.