docs: Add more mongo_sync tutorials.
Added short tutorials for insert and two more for queries.

During this, the mongo_sync_connect example's source was folded into
the generic tut_mongo_sync.c file, together with the rest of the sync
examples.

This is because the examples depend on each other, and putting them
into the same program makes more sense than documenting the order
these programs need to run.