bson: Add bson_cursor_find().
Add a new function: bson_cursor_find(), which takes a cursor and a key
name, and attempts to find the key in the BSON object, starting at the
current cursor position, and wrapping over, if so need be (but only
once).

As a bonus, bson_find() and bson_cursor_find_next() were reimplemented
as thin wrappers around the same _bson_cursor_find() that
bson_cursor_find() uses.