gridfs: Streaming API redesign.
Instead of building on top of mongo_sync_gridfs_file, build directly
on top of mongo_sync_gridfs: the chunk and streaming APIs are two
different ways to access the same files, two incompatible ways.

Therefore, it is pointless to build on onto the other - they're
separate.

As a side effect of the redesign, _stream_from_file() was removed.