tests: Make the test suite optional, and disabled by default.
Due to tar having an arbitrary 99-char file length limit, the test
suite cannot be included in a make dist tarball, unless most of the
files get a rename, which will not happen.

As a workaround, the test suite was made optional, and by default, is
not included in the make dist tarball, which has a side effect of not
even being run by default.

To enable it, use --enable-tests at configure time. But then make dist
will break.