lib: Create a libmongo-client-internal library.
Create a libmongo-client-internal library, which is the same as
libmongo-client, but without symbol versioning.

This allows us to hide internal symbols in the final library, yet,
still write unit tests for them: the tests link with the -internal
library instead.

The header will still need to be hidden, but that can wait, until we
know what exactly we need exported, and what can be hidden completely.