Restore ABI compatibility with 0.1.0.
In 0.1.6, the mongo_connect and mongo_sync_connect symbols were bumped
to LMC_0.1.6, which meant that old programs linked against the
library, that expected to find these symbols at @LMC_0.1.0 suddenly
stopped working.

This was not the intention, so lets provide the @LMC_0.1.0 symbols
along with the @LMC_0.1.6 ones, in such a way that old programs will
exhibit the behaviour expected from them.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>