tests: Only compile libtap when running make check
Compiling libtap when not running the test suite is a waste of
resources, however small, so don't do it. Also adjust the test rules
to follow up on this change.

Furthermore, add a -no-install flag for the tests, so that they won't
get a libtool wrapper (there's no point, they're run from the build
dir only).

Signed-off-by: Gergely Nagy <algernon@balabit.hu>