tests: Make prove optional
Instead of requiring prove to run the test suite, make it possible to
use a different runner. Just set the PROVE variable on the make check
commandline to a script that expects a list of test filenames, relative
to ${top_builddir}, and it will be used instead of prove.

This fixes #21.

Reported-by: Viktor Tusa <tusa@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>